hprose: Move to own pkg
